+++ OGLight/OGLight.meta.js
@@ -1,5 +1,5 @@
 // ==UserScript==
-// @version 4.3.3.5
+// @version 4.3.4
 // @name OGLight
 // @namespace    https://github.com/igoptx/ogameTools
 // ==/UserScript==